+++ src/main/java/hk/org/ha/kcc/its/dto/EamDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class EamDto {
 
-    private Long eamNo; // 1824496
+    private Integer eamNo; // 1824496
     private String SerNo;// E09-123567890123
     private String Model; // ELITDESK PU 8000G5 SFF
     private String BelongTo;// KWH Ward 8A
